Add validation for the manga model
Validation was employed to ensure the data sent to the server from a
manga form is in the appropriate format.

The format of the request body has been modified to ensure that the form
sends arrays of objects in the genre and author fields.

The rendering of the role form group in the form field now utilizes
templates to prevent unnecessary DOM shifts on load. It also ensures
that the roles are linked to a particular author on form submit.