Fix bug in rendering form views
Form views outside of the manga route would crash when attempting to
render a template for author roles, which was only intended for the
manga route. The other routes didn't pass this author data to the form
route, leading to the crash.

This commit ensures that the author template is only rendered in the
manga route.